+++ R/utils.R
@@ -193,32 +193,6 @@ auto_cache_lazy_tensors = function(lts) {
   any(duplicated(map_chr(lts, function(x) dd(x)$dataset_hash)))
 }
 
-clone_graph_unique_ids = function(g) {
-  prev_ids = ids(g$pipeops)
-  pipeops = map(g$pipeops, function(po) po$clone(deep = TRUE))
-
-  new_ids = uniqueify(ids(pipeops), ids(pipeops))
-  walk(seq_along(pipeops), function(i) {
-    pipeops[[i]]$id = new_ids[i]
-  })
-
-  g1 = Graph$new()
-  walk(pipeops, function(po) {
-    g1$add_pipeop(po, clone = FALSE)
-  })
-  new_edges = copy(g$edges)
-
-  id_map = new_ids
-  names(id_map) = prev_ids
-
-  new_edges$src_id = id_map[new_edges$src_id]
-  new_edges$dst_id = id_map[new_edges$dst_id]
-
-  g1$edges = new_edges
-
-  return(g1)
-}
-
 #' Replace the head of a network
 #' Replaces the head of the network with a linear layer with d_out classes.
 #' @param network ([`torch::nn_module`])\cr